netcdfplus cleanup
