+++ openpathsampling/netcdfplus/__init__.py
@@ -1,2 +1,6 @@
 from netcdfplus import NetCDFPlus
-from base import DelayedLoader, lazy_loading_attributes, StorableNamedObject, StorableObject
+from base import StorableNamedObject, StorableObject
+from proxy import DelayedLoader, lazy_loading_attributes, LoaderProxy
+from cache import WeakKeyCache, WeakLRUCache, WeakValueCache, MaxCache, NoCache, Cache, LRUCache
+from json import ObjectJSON, StorableObjectJSON
+from objects import ObjectStore
