+++ openpathsampling/netcdfplus/json.py
