+++ openpathsampling/storage/cv_store.py
@@ -1,4 +1,4 @@
-from openpathsampling.netcdfplus.object_storage import ObjectStore
+from openpathsampling.netcdfplus.objects import ObjectStore
 
 
 class ObjectDictStore(ObjectStore):
